pyproject.toml
[project]
name = "keep_awake"
version = "1.1.0"
description = "Prevent your computer from going to sleep."
readme = "README.md"
keywords = ["sleep", "prevent", "awake", "idle"]
authors = [{ name = "Kevin Chen", email = "1354016594@qq.com" }]
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
]
dependencies = []
optional-dependencies = { linux = ["dbus-python>=1.4.0"] }

[project.urls]
Home-page = "https://github.com/czf0613/keep_awake_py"

[build-system]
requires = ["setuptools>=45.0", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = ["pytest>=7.0.1"]
